requirements.txt
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.0
SQLAlchemy==1.3.5
Werkzeug==0.15.3
python-slugify

